Added HTML fetcher URLConnectionReader.java
You may pass one or more URL as an argument to main (inside the class)
and this program will grab the entirety of the HTML source for that URL
and output it to stdout.
Ex:
java URLConnectionReader http://espn.go.com > espn.html
would make a file named espn.html with all of the html from espn's home
page.

I NEED TO ADD COMMENTS.